refactor: split helpers into focused modules and update docs
